requirements.txt
boto3==1.34.81
botocore==1.34.81
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
urllib3==2.2.1